+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOMigration/MSFT_EXOMigration.psm1
@@ -538,7 +538,6 @@ function Export-TargetResource
 
     try
     {
-        $Script:ExportMode = $true
         [array]$migrationBatches = Get-MigrationBatch -ErrorAction Stop
 
         $i = 1